Remove file filter
The debian/watch file defines rules how to obtain the orig tarball.
The Files-Excluded field automatically filters files from the orig tarball,
which is used, e.g., to get rid of legally problematic or non-source files.

The sources for the filtered files are included in git regardlessly.
This is an impedance mismatch and using origtargz/uscan to get the upstream
source results in build failures because the quilt format checks for untouched
sources. Also, repacking the source should have an effect on the version name,
but it is not represented in this package's version.

In ec80bb75128 ("Add d/missing-sources"),
the sources for the filtered files were added.

Signed-off-by: Bastian Germann <bage@linutronix.de>
Reviewed-by: Torben Hohn <torben.hohn@linutronix.de>