Add a backports profile
Add a pkg.swupdate.bpo profile. With this set, the SWU forwarder handler
and its dependencies are disabled. Without them, the package can be built
on stretch (with debhelper backport) and buster. The problematic packages
are libwebsockets (for both) and curl (for stretch).

Upstream's debian branch disabled that handler with commit
c56f242998f03788 ("debian: temporary drop swuforwarder").

Signed-off-by: Bastian Germann <bastiangermann@fishpost.de>