Move scripts to update from USB Stick to examples
Currently if a USB drive is plugged in it triggers swupdate
to install a provided update from the USB drive.

If signing is deactivated this functionality is a security
risk as arbitrary updates can be installed on the system.

The user should activate the udev rules manually by coping
the necessary files to the correct folders.

Move the udev rules to update a system
with SWUpdate to usr/share/doc/swupdate/examples/udev.

Signed-off-by: Quirin Gylstorff <quirin.gylstorff@siemens.com>