Enable efibootguard by default
As efibootguard is now in the Debian archive, enable building with it by
default. The architecture range is derived from gnu-efi, which is one of
its dependencies.

As there is no need for its build profile any longer replace it with the
bpo profile so the package can still be built for bullseye.