Add swupdate-www
Most gulp plugins of the web-app build systems are not available in
Debian, so build just install the absolutely necessary dependencies.
The general strategy is to have symbolic links to the Debian packages
so we can spare the copyright maintenance.

Font Awesome 5 is not available in Debian, so rely on Fork Awesome which
has a compatibility layer.